Add files via upload
Lights on and off is about Make a program that displays 3 switches and 3 images of turned off lightbulbs, when you click a switch the corresponding lightbulb will change to an image of a turned on lightbulb, and clicking it again will switch it off.

Despedida y saludo  a program that displays 5 switches, if you click a switch, all adjacent switches will follow their opposite state, for example: 
If all switches are off and you turn on switch 3, switches 2 and 4 will turn on as well. After that, if you turn on switch 5, switch 4 will turn off because it is its adjacent.